requirements.txt
aiogram
motor
python-dotenv
certifi
dnspython
